+++ RGBA.Optio.Domain/Models/ResponseModels/MerchantResponseModel.cs
@@ -9,8 +9,8 @@ namespace RGBA.Optio.Domain.Models.ResponseModels
     public class MerchantResponseModel
     {
         public string Name { get; set; }
-        public long Volume {  get; set; }
+        public decimal Volume {  get; set; }
         public long Quantity {  get; set; }
-        public long Average { get; set; }
+        public decimal Average { get; set; }
     }
 }
